Extract find_record into a mixin. Decuplicate.
