+++ app/controllers/mixins/manager_controller_mixin.rb
@@ -4,6 +4,7 @@ module Mixins
 
     included do
       include Mixins::GenericFormMixin
+      include Mixins::FindRecord
     end
 
     def index
@@ -503,20 +504,6 @@ module Mixins
 
     def valid_configured_system_record?(configured_system_record)
       configured_system_record.try(:id)
-    end
-
-    def find_record(model, id)
-      raise _("Invalid input") unless is_integer?(from_cid(id))
-      begin
-        record = Rbac.filtered(model.where(:id => from_cid(id))).first
-      rescue ActiveRecord::RecordNotFound, StandardError => ex
-        if @explorer
-          self.x_node = "root"
-          add_flash(ex.message, :error, true)
-          session[:flash_msgs] = @flash_array.dup
-        end
-      end
-      record
     end
 
     def title
